requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
openai
cryptography
mysql-connector-python

apscheduler

typer

croniter
colorlog

requests
fastmcp
